refactor Manager System
