+++ src/main/java/miniProject_eveningPractice/ManagerSystem.java
@@ -4,7 +4,6 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.groupingBy;
-import static miniProject_eveningPractice.Util.fillMS;
 
 public class ManagerSystem {
 
@@ -13,8 +12,7 @@ public class ManagerSystem {
 
     public void addEmployee(Employee newEmployee) throws EmployeeIdAlreadyExistsException {
         for (Employee employee : employeeList) {
-            if (employee.getId() == newEmployee.getId()) ;
-            {
+            if (employee.getId() == newEmployee.getId()) {
                 throw new EmployeeIdAlreadyExistsException("Employee With Such Id Already Exists");
             }
         }
@@ -103,41 +101,5 @@ public class ManagerSystem {
                     System.out.println("Employee in department " + entry.getKey() + ": " + entry.getValue());
 
                 });
-    }
-
-    public static void main(String[] args) {
-        ManagerSystem managerSystem = fillMS();
-
-        Set<Employee> employeeSet = managerSystem.getEmployeeSet();
-        System.out.println("employeeSet");
-        System.out.println(employeeSet);
-
-        Map<Integer, Employee> employeeByIdMap = managerSystem.getEmployeeByIdMap();
-        System.out.println("employeeByIdMap");
-        System.out.println(employeeByIdMap);
-
-        System.out.println("getEmployeeById");
-        try {
-            System.out.println(managerSystem.getEmployeeById(101));
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-
-        try {
-            System.out.println("displayEmployeesByDepartment");
-            managerSystem.displayEmployeesByDepartment("Logistic");
-            System.out.println("displayEmployeesByJobTitle");
-            managerSystem.displayEmployeesByJobTitle("BackEnd Developer");
-            System.out.println("displayEmployeesByManager");
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-
-        try {
-            managerSystem.displayEmployeesByManager(102);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
-        managerSystem.displayReport();
     }
 }
